Rename scraperTest, indexerBenchmarking and webScraper
